[chore] move adapt and prerender folders
