+++ packages/kit/src/core/adapt/builder.js
@@ -1,5 +1,5 @@
-import { copy, rimraf, mkdirp } from '../../../utils/filesystem.js';
-import { generate_manifest } from '../../../core/generate_manifest/index.js';
+import { copy, rimraf, mkdirp } from '../../utils/filesystem.js';
+import { generate_manifest } from '../generate_manifest/index.js';
 
 /**
  * Creates the Builder which is passed to adapters for building the application.
