+++ packages/kit/src/core/adapt/index.js
