+++ packages/kit/src/core/prerender/fixtures/basic-src/input.html
