+++ packages/kit/src/core/prerender/fixtures/basic-srcset/output.json
