+++ packages/kit/src/core/prerender/fixtures/include-rel-external/output.json
