+++ packages/kit/src/core/prerender/prerender.js
@@ -1,12 +1,12 @@
 import { readFileSync, writeFileSync } from 'fs';
 import { dirname, join } from 'path';
 import { pathToFileURL, URL } from 'url';
-import { mkdirp } from '../../../utils/filesystem.js';
-import { installPolyfills } from '../../../node/polyfills.js';
-import { is_root_relative, normalize_path, resolve } from '../../../utils/url.js';
+import { mkdirp } from '../../utils/filesystem.js';
+import { installPolyfills } from '../../node/polyfills.js';
+import { is_root_relative, normalize_path, resolve } from '../../utils/url.js';
 import { queue } from './queue.js';
 import { crawl } from './crawl.js';
-import { escape_html_attr } from '../../../utils/escape.js';
+import { escape_html_attr } from '../../utils/escape.js';
 
 /**
  * @typedef {import('types').PrerenderErrorHandler} PrerenderErrorHandler
